+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add `AddDefaultIntelliJPlatformDependencies` property to control if default dependencies (`com.intellij` and `intellij.rider`) are added automatically with IntelliJ Platform
 - Use the exact IntelliJ Platform version to resolve platform dependencies when targeting IntelliJ Platform from the nightly channel.
 - Introduce the `intellijPlatformTestRuntimeClasspath` configuration
+- Automatically load the `com.intellij.clion` into the CLion classpath
 
 ### Changed